+++ InstallerLib/Editor.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Xml;
-using System.Collections.Generic;
 using System.ComponentModel;
 using System.Reflection;