+++ UnitTests/dotNetInstallerUnitTests/LoggingUnitTests.cs
@@ -13,15 +13,28 @@ namespace dotNetInstallerUnitTests
     [TestFixture]
     public class LoggingUnitTests
     {
-        [Test]
-        public void TestNoLogging()
+        [SetUp]
+        public void SetUp()
         {
-            Console.WriteLine("TestNoLogging");
-
             if (File.Exists(dotNetInstallerExeUtils.RunOptions.DefaultLogFile))
             {
                 File.Delete(dotNetInstallerExeUtils.RunOptions.DefaultLogFile);
             }
+        }
+
+        [TearDown]
+        public void TearDown()
+        {
+            if (File.Exists(dotNetInstallerExeUtils.RunOptions.DefaultLogFile))
+            {
+                TestContext.AddTestAttachment(dotNetInstallerExeUtils.RunOptions.DefaultLogFile);
+            }
+        }
+
+        [Test]
+        public void TestNoLogging()
+        {
+            Console.WriteLine("TestNoLogging");
 
             ConfigFile configFile = new ConfigFile();
             SetupConfiguration setupConfiguration = new SetupConfiguration();
@@ -41,11 +54,6 @@ namespace dotNetInstallerUnitTests
         public void TestLogConfigSpecified()
         {
             Console.WriteLine("TestLogConfigSpecified");
-
-            if (File.Exists(dotNetInstallerExeUtils.RunOptions.DefaultLogFile))
-            {
-                File.Delete(dotNetInstallerExeUtils.RunOptions.DefaultLogFile);
-            }
 
             ConfigFile configFile = new ConfigFile();
             configFile.log_enabled = true;
@@ -74,11 +82,6 @@ namespace dotNetInstallerUnitTests
         public void TestLogCommandLineOverwritesConfigFile()
         {
             Console.WriteLine("TestLogCommandLineOverwritesConfigFile");
-
-            if (File.Exists(dotNetInstallerExeUtils.RunOptions.DefaultLogFile))
-            {
-                File.Delete(dotNetInstallerExeUtils.RunOptions.DefaultLogFile);
-            }
 
             ConfigFile configFile = new ConfigFile();
             configFile.log_enabled = true;
@@ -109,11 +112,6 @@ namespace dotNetInstallerUnitTests
             Console.WriteLine("TestLogAcceptsPathVariables");
 
             string resolved_logfile = Path.Combine(Path.GetTempPath(), "TestLogAcceptsPathVariables.log");
-            if (File.Exists(dotNetInstallerExeUtils.RunOptions.DefaultLogFile))
-            {
-                File.Delete(dotNetInstallerExeUtils.RunOptions.DefaultLogFile);
-            }
-
             if (File.Exists(resolved_logfile))
             {
                 File.Delete(resolved_logfile);