+++ UnitTests/dotNetUnitTestsRunner/InstallerLinkerExeUtils.cs
@@ -1,12 +1,10 @@
 using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Web;
 using System.IO;
 using System.Reflection;
 using NUnit.Framework;
 using System.Diagnostics;
-using System.Threading;
 using InstallerLib;
 
 namespace dotNetUnitTestsRunner
@@ -72,7 +70,7 @@ namespace dotNetUnitTestsRunner
                     // locate dotnetinstaller.exe
                     Uri uri = new Uri(Assembly.GetExecutingAssembly().CodeBase);
                     _location = Path.GetFullPath(Path.Combine(Path.GetDirectoryName(HttpUtility.UrlDecode(uri.AbsolutePath)),
-                        string.Format(@"..\..\..\..\InstallerLinker\bin\{0}\", configuration)));
+                        string.Format(@"..\..\..\..\..\InstallerLinker\bin\{0}\net40\", configuration)));
                 }
 
                 return _location;